Improve on tests

+++ controller/include/controller.h
@@ -6,16 +6,16 @@
 class Controller : public IController
 {
 public:
-    Controller(uint64_t intergral_gain, uint64_t derivative_gain, uint64_t proportional_gain);
+    Controller(double intergral_gain, double derivative_gain, double proportional_gain);
     ~Controller();
 
     ParamType GetParams() override;
 
 private:
-    uint64_t error_{};
-    uint64_t intergral_gain_{};
-    uint64_t derivative_gain_{};
-    uint64_t proportional_gain_{};
+    double error_{};
+    double intergral_gain_{};
+    double derivative_gain_{};
+    double proportional_gain_{};
 };
 
 #endif /* CONTROLLER_CONTROLLER_H */
